refactor: Split and reorganize tests
