+++ infinitest-runner/src/main/java/org/infinitest/testrunner/junit5/Junit5Runner.java
@@ -27,30 +27,15 @@
  */
 package org.infinitest.testrunner.junit5;
 
-import static org.infinitest.testrunner.TestEvent.methodFailed;
 import static org.junit.platform.engine.discovery.DiscoverySelectors.selectClass;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
 import org.infinitest.config.InfinitestConfigurationSource;
-import org.infinitest.testrunner.Clock;
-import org.infinitest.testrunner.MethodStats;
-import org.infinitest.testrunner.SystemClock;
-import org.infinitest.testrunner.TestEvent;
 import org.infinitest.testrunner.TestResults;
 import org.junit.jupiter.engine.JupiterTestEngine;
-import org.junit.platform.commons.util.PreconditionViolationException;
-import org.junit.platform.engine.TestExecutionResult;
-import org.junit.platform.engine.TestSource;
-import org.junit.platform.engine.support.descriptor.MethodSource;
 import org.junit.platform.launcher.EngineFilter;
 import org.junit.platform.launcher.Launcher;
 import org.junit.platform.launcher.LauncherDiscoveryRequest;
 import org.junit.platform.launcher.TestExecutionListener;
-import org.junit.platform.launcher.TestIdentifier;
 import org.junit.platform.launcher.TestPlan;
 import org.junit.platform.launcher.core.LauncherDiscoveryRequestBuilder;
 import org.junit.platform.launcher.core.LauncherFactory;
@@ -76,97 +61,6 @@ public class Junit5Runner {
 		launcher.execute(request, new TestExecutionListener[0]);
 
 		return listener.getTestResults();
-	}
-
-	private static class JUnit5EventTranslator implements TestExecutionListener {
-		private final List<TestEvent> eventsCollected;
-		private final Map<TestIdentifier, MethodStats> methodStats;
-		private final Clock clock;
-
-		public JUnit5EventTranslator() {
-			this.clock = new SystemClock();
-			eventsCollected = new ArrayList<TestEvent>();
-			methodStats = new HashMap<TestIdentifier, MethodStats>();
-		}
-
-		@Override
-		public void executionStarted(TestIdentifier testIdentifier) {
-			if (testIdentifier.isTest()) {
-				getMethodStats(testIdentifier).startTime = clock.currentTimeMillis();
-			}
-		}
-
-		@Override
-		public void executionFinished(TestIdentifier testIdentifier, TestExecutionResult testExecutionResult) {
-			if (testIdentifier.isTest()) {
-				getMethodStats(testIdentifier).stopTime = clock.currentTimeMillis();
-				switch (testExecutionResult.getStatus()) {
-
-				case SUCCESSFUL:
-					break;
-
-				case ABORTED: {
-					break;
-				}
-
-				case FAILED: {
-					eventsCollected.add(createEventFrom(testIdentifier, testExecutionResult));
-					break;
-				}
-
-				default:
-					throw new PreconditionViolationException(
-							"Unsupported execution status:" + testExecutionResult.getStatus());
-				}
-			}
-		}
-
-		private TestEvent createEventFrom(TestIdentifier testIdentifier, TestExecutionResult testExecutionResult) {
-			String testCaseName = getTestCaseName(testIdentifier);
-			Throwable exception = testExecutionResult.getThrowable().get();
-
-			return methodFailed(testCaseName, getMethodName(testIdentifier), exception);
-		}
-
-		private static String getTestCaseName(TestIdentifier description) {
-			try {
-				TestSource testSource = description.getSource().get();
-				if (testSource instanceof MethodSource) {
-					return ((MethodSource) testSource).getClassName();
-				}
-				return description.getDisplayName().split("\\(|\\)")[1];
-			} catch (Exception e) {
-				return description.getDisplayName();
-			}
-		}
-
-		public TestResults getTestResults() {
-			TestResults results = new TestResults(eventsCollected);
-			results.addMethodStats(methodStats.values());
-			return results;
-		}
-
-		private String getMethodName(TestIdentifier description) {
-			try {
-				TestSource testSource = description.getSource().get();
-				if (testSource instanceof MethodSource) {
-					return ((MethodSource) testSource).getMethodName();
-				}
-				return description.getDisplayName().split("\\(|\\)")[0];
-			} catch (Exception e) {
-				return description.getDisplayName();
-			}
-		}
-
-		private MethodStats getMethodStats(TestIdentifier description) {
-			MethodStats stats = methodStats.get(description);
-			if (stats == null) {
-				stats = new MethodStats(getMethodName(description));
-				methodStats.put(description, stats);
-			}
-			return stats;
-		}
-
 	}
 
 	public static boolean isJUnit5Test(Class<?> clazz) {
