+++ infinitest-runner/src/test/java/org/infinitest/testrunner/exampletests/testng/TestWithTestNGGroupsAndSetup.java
@@ -25,20 +25,27 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.infinitest.testrunner;
+package org.infinitest.testrunner.exampletests.testng;
 
-import org.testng.*;
+import static org.testng.Assert.*;
+
 import org.testng.annotations.*;
 
-@Test
-public class TestWithTestNGClassAnnotationOnly {
-	static boolean fail;
+public class TestWithTestNGGroupsAndSetup {
+	public static boolean setupWasCalled;
 
-	public void shouldPass() {
-		Assert.assertTrue(true);
+	@BeforeMethod(groups = { "automated", "integration" })
+	public void setup() {
+		setupWasCalled = true;
 	}
 
-	public void shouldFail() {
-		Assert.assertFalse(fail);
+	@Test(groups = "automated")
+	public void automated() {
+		assertFalse(true);
+	}
+
+	@Test(groups = "integration")
+	public void integration() {
+		assertFalse(true);
 	}
 }
