+++ pom.xml
@@ -196,7 +196,8 @@
 					<excludes>
 						<exclude>**/*$*</exclude>
 						<exclude>com/fakeco/**</exclude>
-						<exclude>**/NonPublicJUnit4Test*</exclude>
+						
+						<exclude>**/exampletests/**</exclude>
 					</excludes>
 					<!-- argLine is configured using a property to simplify jacoco configuration -->
 				</configuration>
